clean: remove async methods
When using async functions, babel requires the "regeneratorRuntime"
included (which is available either via `babel-polyfill` or
`babel-plugin-transform-runtime`). A library should not expect the
consumer to provide such a function (because babel may not be used at
all in the final project). On top of that, the async/await functions
can be trivialy changed to simple functions by removing some code.